backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
statsmodels
xgboost
apscheduler
requests
python-dotenv
sqlalchemy
psycopg2-binary
joblib